moved case class for status into separate package
