+++ pom.xml
@@ -59,6 +59,7 @@
         <confluent.version>2.0.1</confluent.version>
         <kafka.version>0.9.0.1</kafka.version>
         <hosebird.version>2.2.0</hosebird.version>
+        <avro.version>1.7.7</avro.version>
         <jsr305.version>1.3.9</jsr305.version>
         <mockito.version>1.10.19</mockito.version>
         <scalatest.version>3.0.0-M1</scalatest.version>
@@ -126,7 +127,7 @@
         <dependency>
             <groupId>org.apache.avro</groupId>
             <artifactId>avro</artifactId>
-            <version>1.8.0</version>
+            <version>${avro.version}</version>
         </dependency>
     </dependencies>
